refactor: 예약 도메인 조회 시 예약 시간을 참조하도록 변경
