+++ src/main/java/roomescape/service/DefaultReservationService.java
@@ -5,15 +5,18 @@ import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Profile;
 import roomescape.domain.Reservation;
-import roomescape.dto.ReservationRequestDTO.AddReservationRequest;
-import roomescape.dto.ReservationResponseDTO.AddReservationResponse;
-import roomescape.dto.ReservationResponseDTO.QueryReservationResponse;
+import roomescape.domain.Time;
+import roomescape.dto.reservation.ReservationRequestDTO.AddReservationRequest;
+import roomescape.dto.reservation.ReservationResponseDTO.AddReservationResponse;
+import roomescape.dto.reservation.ReservationResponseDTO.QueryReservationResponse;
 import roomescape.repository.ReservationRepository;
+import roomescape.repository.TimeRepository;
 
 @Profile("default")
 @RequiredArgsConstructor
 public class DefaultReservationService implements ReservationService {
 	private final ReservationRepository reservationRepository;
+	private final TimeRepository timeRepository;
 
 	@Override
 	public List<QueryReservationResponse> getReservations() {
@@ -22,19 +25,20 @@ public class DefaultReservationService implements ReservationService {
 						reservation.id(),
 						reservation.name(),
 						reservation.date(),
-						reservation.time()))
+						reservation.time().time_value()))
 				.collect(Collectors.toList());
 	}
 
 	@Override
 	public AddReservationResponse addReservation(
 			AddReservationRequest reservationRequest) {
+		Time time = timeRepository.findById(reservationRequest.time_id());
 
 		Reservation newReservation = new Reservation(
 				null,
 				reservationRequest.name(),
 				reservationRequest.date(),
-				reservationRequest.time()
+				time
 		);
 
 		Long savedReservationId = reservationRepository.save(newReservation);
@@ -43,7 +47,8 @@ public class DefaultReservationService implements ReservationService {
 				savedReservationId,
 				newReservation.name(),
 				newReservation.date(),
-				newReservation.time());
+				newReservation.time().time_value()
+		);
 	}
 
 	@Override
